step 2 cleanup
