--- app.log
@@ -1,6 +1,0 @@
-2024-12-28 18:09:32,439 - INFO - Python logging initialized
-2024-12-28 18:09:35,292 - INFO - 127.0.0.1 - - [2024-12-28 18:09:35] "GET /main3.html HTTP/1.1" 404 881 0.003000
-2024-12-28 18:09:35,408 - INFO - 127.0.0.1 - - [2024-12-28 18:09:35] "GET /favicon.ico HTTP/1.1" 404 882 0.001003
-2024-12-28 18:10:49,975 - INFO - Python logging initialized
-2024-12-28 18:10:51,565 - INFO - 127.0.0.1 - - [2024-12-28 18:10:51] "GET /main3.html HTTP/1.1" 404 881 0.002982
-2024-12-28 18:10:51,620 - INFO - 127.0.0.1 - - [2024-12-28 18:10:51] "GET /favicon.ico HTTP/1.1" 404 882 0.001000
