+++ backend/cy_cookies.py
@@ -7,12 +7,12 @@ import os
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-cookies = Blueprint('cookies', __name__)
+cy_cookies = Blueprint('cy_cookies', __name__)
 need_reload=True
 # Define the path to the JSON file
 json_file_path = 'cookies.json'
 
-@cookies.route('/save_cookie', methods=['POST'])
+@cy_cookies.route('/save_cookie', methods=['POST'])
 def save_cookie():
     global cookies_data
     cookie_value = request.json.get('cookie_value')
@@ -39,7 +39,7 @@ def save_cookie():
     
     return jsonify({"message": "done"})
 
-@cookies.route('/load_cookies', methods=['GET'])
+@cy_cookies.route('/load_cookies', methods=['GET'])
 def load_cookies():
     global cookies_data
     if os.path.exists(json_file_path):
@@ -51,7 +51,7 @@ def load_cookies():
         #logger.warning("JSON file does not exist.")
         return jsonify({"error": "JSON file does not exist"}), 404
 
-@cookies.route('/get_cookie', methods=['POST'])
+@cy_cookies.route('/get_cookie', methods=['POST'])
 def get_cookie():
     cookie_name = request.json.get('cookie_name')
     if cookie_name is None:
@@ -75,10 +75,19 @@ def get_cookie_value(key_name):
     try:
         if key_name is None:
             return jsonify({"error": "key_name is required"}), 400
+        
+        # Initialiser cookies_data en lisant le fichier
+        if os.path.exists(json_file_path):
+            with open(json_file_path, 'r') as file:
+                cookies_data = json.load(file)
+        else:
+            cookies_data = {}
+            
         # Get the value for the provided key
         value = cookies_data.get(key_name)
         #logger.info(f"dbg5643 Key: {key_name}, Value: {value}")
         return value
     except Exception as e:
         logger.error(f"Error-3345 getting cookie value for key {key_name}: {e}")
+        return None
