+++ static/js/cy_all_dos_filters.js
@@ -1,0 +1,53 @@
+function updateFilterValues(filters) {
+    //console.log('Filters object:', typeof(filters), filters); // Vérifier le format de filters
+    // Itérer directement sur les clés de l'objet filters
+
+    // forEach((col =W
+    for (const key in filters) {
+        if (filters.hasOwnProperty(key)) {
+            //console.log(`Filter value for ${key}:`, filters[key]); 
+            const filterElement = document.getElementById(`filter-${key}`);
+            const selectElement = document.getElementById(`select-${key}`);
+            if (filterElement) {
+                filterElement.value = filters[key] || '';
+                //console.log(`Filter value for ${key}:`, filterElement.value); // Add console.log
+            }
+            if (selectElement) {
+                const option = document.createElement('option');
+                option.value = filters[key];
+                option.textContent = filters[key];
+                selectElement.appendChild(option);
+                selectElement.value = filters[key];
+            }
+        }
+    }
+}
+
+function saveFilterValues(filters) {
+    fetch('/save_filters_json', {
+        method: 'POST',
+        headers: {
+            'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({
+            filters: filters,
+            tabActive: window.tabActive
+        })
+    })
+    .then(response => {
+        if (!response.ok) {
+            throw new Error('Network response was not ok');
+        }
+        return response.json();
+    })
+    .then(data => {
+        if (data.status === "success") {
+            //console.log('Filter values successfully saved.');
+        } else {
+            console.error('Error saving filter values:', data.message);
+        }
+    })
+    .catch(error => {
+        console.error('An unexpected error occurred while saving filter values:', error);
+    });
+}
