+++ static/js/cy_main.js
@@ -116,7 +116,7 @@ function save_config_col() {
         console.error('Error saving configuration:', error);
     });
 }
-//save_config_col();
+
 
 
 function  view_results()
@@ -748,196 +748,13 @@ function saveTableData() {
     });
 }
 
-function generate_index_html()
-{
-    fetch('/generate_html_index', {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({ dossier_list: window.annonces, sufix: window.CONSTANTS['FILE_NAMES']['ANNONCE_SUFFIX'] })
-
-    })
-    .then(response => response.json())
-    .then(data => {
-        if (data.status === "success") {
-            console.log('dbg4757 Data successfully saved.');
-        } else {
-            console.error('Error saving data:', data.message);
-        }
-    })
-    .catch(error => {
-        console.error('Error saving data:', error);
-    });
-}
-
-function saveFilterValues(filters) {
-    fetch('/save_filters_json', {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({
-            filters: filters,
-            tabActive: window.tabActive
-        })
-    })
-    .then(response => {
-        if (!response.ok) {
-            throw new Error('Network response was not ok');
-        }
-        return response.json();
-    })
-    .then(data => {
-        if (data.status === "success") {
-            //console.log('Filter values successfully saved.');
-        } else {
-            console.error('Error saving filter values:', data.message);
-        }
-    })
-    .catch(error => {
-        console.error('An unexpected error occurred while saving filter values:', error);
-    });
-}
 
 
-function updateFilterValues(filters) {
-    //console.log('Filters object:', typeof(filters), filters); // Vérifier le format de filters
-    // Itérer directement sur les clés de l'objet filters
 
-    // forEach((col =W
-    for (const key in filters) {
-        if (filters.hasOwnProperty(key)) {
-            //console.log(`Filter value for ${key}:`, filters[key]); 
-            const filterElement = document.getElementById(`filter-${key}`);
-            const selectElement = document.getElementById(`select-${key}`);
-            if (filterElement) {
-                filterElement.value = filters[key] || '';
-                //console.log(`Filter value for ${key}:`, filterElement.value); // Add console.log
-            }
-            if (selectElement) {
-                const option = document.createElement('option');
-                option.value = filters[key];
-                option.textContent = filters[key];
-                selectElement.appendChild(option);
-                selectElement.value = filters[key];
-            }
-        }
-    }
-}
+
+
 
 // Function to load filter values from JSON file and update input fields
-
-//onfiche
-function openEditModal(rowId) {
-    // Find the annonce data
-    const index = window.annonces.findIndex(a => Object.keys(a)[0] === rowId);
-    if (index === -1) return;
-    const annonce = window.annonces[index][rowId];
-
-    // Définir les groupes d'onglets
-    const tabGroups = {
-        'Informations principales': ['dossier', 'description', 'id', 'entreprise', 'categorie'],
-        'Statut': ['etat', 'lien_Etape','annonce_pdf','CV','CVfile'],
-        'suivi': [ 'Date','Date_rep','todo','commetaires','Origine'],
-        'Contact': [ 'contact','tel', 'mail','url'],
-        'Détails': ['Commentaire', 'type', 'Lieux'], 
-        'GPT': ['GptSum', 'instructions','request'],
-        'Publication': ['lnk_Youtub_value','path_dirpartage_value']
-    };
-
-    // Create modal HTML with tabs
-    let modalHtml = `
-        <dialog id="editModal" class="edit-modal">
-            <form method="dialog">
-                <h2>Modifier l'annonce</h2>
-                <div class="edit-tabs">
-                    ${Object.keys(tabGroups).map((tabName, index) => `
-                        <button type="button" class="tab-button ${index === 0 ? 'active' : ''}" 
-                                onclick="switchEditTab(event, '${tabName}')">${tabName}</button>
-                    `).join('')}
-                </div>
-
-                ${Object.entries(tabGroups).map(([tabName, fields], index) => `
-                    <div id="${tabName}" class="tab-content ${index === 0 ? 'active' : ''}">
-                        ${fields.map(field => `
-                            <div class="form-group">
-                                <label>${field}:</label>
-                                ${field === 'instructions' ? 
-                                    `<textarea id="edit-${field}" rows="5">${annonce[field] || ''}</textarea>` :
-                                    `<input type="text" id="edit-${field}" value="${annonce[field] || ''}" 
-                                       ${field === 'dossier' ? 'readonly' : ''}>`
-                                }
-                            </div>
-                        `).join('')}
-                    </div>
-                `).join('')}
-
-                <div class="button-group">
-                    <button type="button" onclick="saveEdit('${rowId}')">Enregistrer</button>
-                    <button type="button" onclick="cancelEdit()">Annuler</button>
-                </div>
-            </form>
-        </dialog>
-    `;
-
-    // Remove existing modal if any
-    const existingModal = document.getElementById('editModal');
-    if (existingModal) {
-        existingModal.remove();
-    }
-
-    // Add modal to document
-    document.body.insertAdjacentHTML('beforeend', modalHtml);
-
-    // Show modal
-    const modal = document.getElementById('editModal');
-    modal.showModal();
-}
-
-function switchEditTab(event, tabName) {
-    // Hide all tab contents
-    const tabContents = document.getElementsByClassName('tab-content');
-    for (let content of tabContents) {
-        content.classList.remove('active');
-    }
-
-    // Remove active class from all tab buttons
-    const tabButtons = document.getElementsByClassName('tab-button');
-    for (let button of tabButtons) {
-        button.classList.remove('active');
-    }
-
-    // Show the selected tab content and activate the button
-    document.getElementById(tabName).classList.add('active');
-    event.currentTarget.classList.add('active');
-}
-
-function saveEdit(rowId) {
-    // Ensure the index is correct
-    const index = window.annonces.findIndex(a => Object.keys(a)[0] === rowId);
-    if (index === -1) return;
-    const annonce = window.annonces[index][rowId];
-
-    // Update annonce with new values based on window.columns
-    window.columns.forEach(col => {
-        const inputElement = document.getElementById(`edit-${col.key}`);
-        if (inputElement) {
-            annonce[col.key] = inputElement.value;
-        }
-    });
-
-    // Save changes and refresh table
-    refresh();
-  
-
-    // Close modal
-    document.getElementById('editModal').close();
-}
-
-function cancelEdit() {
-    document.getElementById('editModal').close();
-}
 
 
 
@@ -957,151 +774,8 @@ function set_current_row() {
     }
 }
 
-function toggleHeaderVisibility() {
-    const header = document.querySelector('thead');
-    const toggleButton = document.getElementById('toggle-header-btn');
-    if (header.style.display === 'none') {
-        header.style.display = '';
-        toggleButton.textContent = 'Hide Header';
-    } else {
-        header.style.display = 'none';
-        toggleButton.textContent = 'Show Header';
-    }
-}
 
-function generateTableHeaders() {
-    const thead = document.querySelector('thead tr');
-    const filterRow = document.querySelector('thead tr:nth-child(2)');
-    // Clear existing headers
-    thead.innerHTML = '';
-    filterRow.innerHTML = '';
-    
-    window.columns.forEach(col => {
-        if (colisvisible(col.type) && col.visible === true) {
-            // Create header cell
-            const th = document.createElement('th');
-            th.style.width = col.width;
-            th.classList.add('filter-cell-col');
-            //th.textContent = col.title.charAt(0).toUpperCase() + col.key.slice(1);
-            th.textContent = col.title;
-            thead.appendChild(th);
 
-            // Create filter cell
-            const filterCell = document.createElement('th');
-            
-            filterCell.classList.add('filter-cell');
-            if (col.key === 'categorie' || col.key === 'etat' || col.key === 'todo' || col.key === 'type') {
-                const container = document.createElement('div');
-                container.classList.add('filter-container');
-
-                const input = document.createElement('input');
-                input.type = 'text';
-                input.id = `filter-${col.key}`;
-                input.placeholder = 'Filter';
-                container.appendChild(input);
-
-                const select = document.createElement('select');
-                select.id = `select-${col.key}`;
-                const option = document.createElement('option');
-                option.value = '';
-                option.textContent = 'All';
-                select.appendChild(option);
-                container.appendChild(select);
-
-                filterCell.appendChild(container);
-
-                input.addEventListener('input', filterTable);
-                select.addEventListener('change', () => {
-                    input.value = select.value;
-                    filterTable();
-                });
-            } else {
-                const input = document.createElement('input');
-                input.classList.add('filter-container')
-                input.type = 'text';
-                input.id = `filter-${col.key}`;
-                input.placeholder = 'Filter';
-                filterCell.appendChild(input);
-                input.addEventListener('input', filterTable);
-            }
-            filterCell.classList.add('filter-cell');
-            filterRow.appendChild(filterCell);
-        }
-    });
-
-    // Add status header and filter cell
-    const statusHeader = document.createElement('th');
-    statusHeader.classList.add('header');
-    // statusHeader.textContent = 'Status';
-    thead.appendChild(statusHeader);
-
-    const statusFilterCell = document.createElement('th');
-    
-}
-
-function setNewTab(){
-    const activeTab = document.querySelector('.tab.active');
-    if (activeTab !== null) {
-        window.tabActive=activeTab.textContent;
-    }
-}
-
-function changeTab(tabName) {
-    // Remove 'active' class from all tabs
-    document.querySelectorAll('.tab').forEach(tab => {
-        tab.classList.remove('active');
-    });
-    // Add 'active' class to the clicked tab
-    document.querySelector(`.tab[onclick="changeTab('${tabName}')"]`).classList.add('active');
-  
-    setNewTab();
-    refresh();
-}
-
-// Function to generate the column visibility form
-function generateColumnVisibilityForm() {
-    const formHtml = `
-        <dialog id="columnVisibilityModal" class="modal">
-            <form id="columnVisibilityForm">
-                <h3>Manage Column Visibility</h3>
-                ${window.columns.map((col, index) => `
-                    <div>
-                        <input type="checkbox" id="col-${index}" ${col.visible ? 'checked' : ''}>
-                        <label for="col-${index}">${col.key.charAt(0).toUpperCase() + col.key.slice(1)}</label>
-                    </div>
-                `).join('')}
-                <button type="button" onclick="saveColumnVisibility()">Save</button>
-                <button type="button" onclick="toggleColumnVisibilityForm()">Close</button>
-            </form>
-        </dialog>
-    `;
-    document.body.insertAdjacentHTML('beforeend', formHtml);
-}
-
-// Function to save column visibility settings
-function saveColumnVisibility() {
-    window.columns.forEach((col, index) => {
-        const checkbox = document.getElementById(`col-${index}`);
-        if (checkbox) {
-            col.visible = checkbox.checked;
-        }
-    });
-    save_config_col(window.columns);
-    
-    toggleColumnVisibilityForm(); // Close the modal after saving
-}
-
-// Function to toggle the visibility of the column visibility form
-function toggleColumnVisibilityForm() {
-    const modal = document.getElementById('columnVisibilityModal');
-    if (modal) {
-        if (modal.open) {
-            modal.close();
-        } else {
-            modal.showModal();
-        }
-    }
-}
 async function loadCookies() {
         try {
             const response = await fetch('/load_cookies', {
@@ -1185,39 +859,6 @@ function serializeColumns(columns) {
 
 
 
-function loadColumnsFromJson() {
-    fetch('/load_config_col', {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({ tabActive: window.tabActive })
-    })
-    .then(response => response.json())
-    .then(data => {
-        window.columns = deserializeColumns(data);
-        reassignEventHandlers(window.columns);
-        generateTableHeaders();
-        generateColumnVisibilityForm();
-    })
-    .catch(error => {
-        console.error('Error loading column configuration:', error);
-    });
-}
-// Function to reassign event handlers after deserialization
-function reassignEventHandlers(columns) {
-    columns.forEach(col => {
-        if (col.eventHandler === 'openUrlHandler') {
-            col.eventHandler = openUrlHandler;
-        } else if (col.eventHandler === 'SelectHandler') {
-            col.eventHandler = SelectHandler;
-        }
-        
-        // Add more handlers as needed
-    });
-}
-
-
 // Function to deserialize columns
 function deserializeColumns(columns) {
     return columns.map(col => {
@@ -1258,98 +899,6 @@ function open_dir(filepath) {
         console.error('Error opening directory:', error);
     });
 }
-
-
-function getFilePathFromRowId(rowId) {
-    // Implémentez cette fonction pour obtenir le chemin du fichier à partir de l'ID de la ligne
-    // Par exemple :
-    return document.querySelector(`#row-${rowId}`).dataset.filePath;
-}
-
-function AIQ(ispdf, value, oneitem) {
-    const params = new URLSearchParams();
-    params.append('ispdf', ispdf);
-    params.append('value', encodeURIComponent(value));
-    params.append('description', encodeURIComponent(oneitem.description));
-   
-    window.open(`qa.html?${params.toString()}`, '_blank');
-}
-
-// ...existing code...
-
-function loadCSVFile(filePath) {
-    fetch('/read_csv_file', {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({ file_path: filePath })
-    })
-    .then(response => response.json())
-    .then(data => {
-        displayCSVTable(data);
-    })
-    .catch(error => {
-        console.error('Error reading CSV file:', error);
-    });
-}
-
-function displayCSVTable(data) {
-    const tableContainer = document.getElementById('csv-table-container');
-    if (!tableContainer) {
-        console.error('Element with id "csv-table-container" not found.');
-        return;
-    }
-
-    const table = document.createElement('table');
-    table.classList.add('csv-table');
-
-    // Create table header
-    const thead = document.createElement('thead');
-    const headerRow = document.createElement('tr');
-    if (data.length > 0) {
-        Object.keys(data[0]).forEach(key => {
-            const th = document.createElement('th');
-            th.textContent = key;
-            headerRow.appendChild(th);
-        });
-    }
-    thead.appendChild(headerRow);
-    table.appendChild(thead);
-
-    // Create table body
-    const tbody = document.createElement('tbody');
-    data.forEach(row => {
-        const tr = document.createElement('tr');
-        Object.values(row).forEach(value => {
-            const td = document.createElement('td');
-            td.textContent = value;
-            tr.appendChild(td);
-        });
-        tbody.appendChild(tr);
-    });
-    table.appendChild(tbody);
-
-    // Clear existing content and append the new table
-    tableContainer.innerHTML = '';
-    tableContainer.appendChild(table);
-}
-
-function openCSVTable(filePath) {
-    const params = new URLSearchParams();
-    params.append('file', filePath);
-    window.open(`csv_table.html?${params.toString()}`, '_blank');
-}
-
-/* // Add menu
-function createMenu() {
-    const menu = document.createElement('div');
-    menu.id = 'menu';
-    menu.innerHTML = `
-        <button onclick="openCSVTable('Suivi_annonce_apec.csv')">Load CSV File</button>
-    `;
-    document.body.insertBefore(menu, document.body.firstChild);
-} */
