+++ templates/index.html
@@ -97,10 +97,19 @@
       </ul>
     </div>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="{{ url_for('static', filename='js/RP_main.js') }}"></script>
-    <script src="{{ url_for('static', filename='js/RP_notes.js') }}"></script>
-    <script src="{{ url_for('static', filename='js/RP_selectRep.js') }}"></script>  
-    <script src="{{ url_for('static', filename='js/RP_conf.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/cy_main.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/cy_one_dos_notes.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/cy_one_dos_Dir.js') }}"></script>  
+    <script src="{{ url_for('static', filename='js/cy_one_dos_edit.js') }}"></script>
+
+    <script src="{{ url_for('static', filename='js/cy_all_dos_tabs.js') }}"></script>
+
+    <script src="{{ url_for('static', filename='js/cy_all_dos_index.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/cy_all_dos_header.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/cy_all_dos_filters.js') }}"></script>
+ 
+    
+   
     
   </body>
 </html>
